Update zip functionality

+++ server/controllers/assessments.controller.js
@@ -271,21 +271,25 @@ function getAssets(query) {
 
 function getAllDataOnUserId(userId, res) {
   deleteZippedForIdIfExists(userId)
-  res.attachment(userId + '.zip')
   const folderPath = path.join('assessment_data', userId)
-  const output = fs.createWriteStream(path.join(folderPath, userId + '.zip'))
-  const archive = archiver('zip', {})
-  output.on('close', function () {
-    console.log(archive.pointer() + ' total bytes')
-    console.log('archiver has been finalized and the output file descriptor has closed.')
+  if (fs.existsSync(folderPath)) {
+    res.attachment(userId + '.zip')
+    const output = fs.createWriteStream(path.join(folderPath, userId + '.zip'))
+    const archive = archiver('zip', {})
+    output.on('close', function () {
+      console.log(archive.pointer() + ' total bytes')
+      console.log('archiver has been finalized and the output file descriptor has closed.')
+      return res.status(200).send('OK').end()
+      // KRM: Return promise here!
+    })
+    const directory = path.join('assessment_data', userId)
+    archive.directory(directory, false)
+    archive.pipe(res)
+    archive.finalize()
     deleteZippedForIdIfExists(userId)
-    return res.status(200).send('OK').end()
-    // KRM: Return promise here!
-  })
-  const directory = path.join('assessment_data', userId)
-  archive.directory(directory, false)
-  archive.pipe(res)
-  archive.finalize()
+  } else {
+    res.sendStatus(404)
+  }
 }
 
 function getAudioDataForHashKey(hashKey) {
@@ -296,9 +300,7 @@ function deleteZippedForIdIfExists(userId) {
   const deleteFile = path.join('assessment_data', userId, userId + '.zip')
   if (fs.existsSync(deleteFile)) {
     console.log('delete')
-    fs.unlinkSync(deleteFile, (err) => {
-      if (err) console.log(err)
-    })
+    fs.unlinkSync(deleteFile)
   }
 }
